refactor: change directory to _components
